+++ src/modules/admin/admin.controller.ts
@@ -11,6 +11,7 @@ import {
   ApiBearerAuth,
   ApiHeader,
   ApiOperation,
+  ApiSecurity,
   ApiTags,
 } from "@nestjs/swagger";
 
@@ -20,6 +21,7 @@ import { StatusService } from "../status/status.service";
 import { Role } from "../users/models/role.enum";
 
 @ApiBearerAuth()
+@ApiSecurity('apikey')
 @Controller("admin")
 @ApiTags("admin")
 export class AdminController {