+++ src/modules/auth/controllers/authentication.controller.ts
@@ -13,6 +13,7 @@ import {
   ApiBody,
   ApiOkResponse,
   ApiOperation,
+  ApiSecurity,
   ApiTags,
 } from "@nestjs/swagger";
 import { SkipGuards } from "../../../decorators/skip-guards.decorator";
@@ -26,6 +27,7 @@ import { Session } from "../session.entity";
 @Controller("auth")
 @ApiTags("auth")
 @ApiBearerAuth()
+@ApiSecurity('apikey')
 export class GamevaultJwtController {
   private readonly logger = new Logger(this.constructor.name);
   constructor(private readonly authService: AuthenticationService) {}