+++ src/modules/games/games.controller.ts
@@ -17,6 +17,7 @@ import {
   ApiHeader,
   ApiOkResponse,
   ApiOperation,
+  ApiSecurity,
   ApiTags,
 } from "@nestjs/swagger";
 import { InjectRepository } from "@nestjs/typeorm";
@@ -49,6 +50,7 @@ import { UpdateGameDto } from "./models/update-game.dto";
 @ApiBearerAuth()
 @ApiTags("game")
 @Controller("games")
+@ApiSecurity('apikey')
 export class GamesController {
   private readonly logger = new Logger(this.constructor.name);