+++ src/modules/metadata/metadata.controller.ts
@@ -4,6 +4,7 @@ import {
   ApiOkResponse,
   ApiOperation,
   ApiQuery,
+  ApiSecurity,
   ApiTags,
 } from "@nestjs/swagger";
 
@@ -17,6 +18,7 @@ import { ProviderSlugDto } from "./providers/models/provider-slug.dto";
 @Controller("metadata")
 @ApiTags("metadata")
 @ApiBearerAuth()
+@ApiSecurity('apikey')
 export class MetadataController {
   constructor(private readonly metadataService: MetadataService) {}