+++ src/modules/users/users.module.ts
@@ -4,9 +4,8 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import { GamesModule } from "../games/games.module";
 import { MediaModule } from "../media/media.module";
 import { ActivityGateway } from "./activity.gateway";
+import { ApiKeyService } from "./api-key.service";
 import { GamevaultUser } from "./gamevault-user.entity";
-import { ApiKeyGuard } from "./socket-secret.guard";
-import { ApiKeyService } from "./socket-secret.service";
 import { UsersController } from "./users.controller";
 import { UsersService } from "./users.service";
 
@@ -17,7 +16,7 @@ import { UsersService } from "./users.service";
     forwardRef(() => GamesModule),
   ],
   controllers: [UsersController],
-  providers: [UsersService, ApiKeyService, ApiKeyGuard, ActivityGateway],
-  exports: [UsersService],
+  providers: [UsersService, ApiKeyService, ActivityGateway],
+  exports: [UsersService, ApiKeyService],
 })
 export class UsersModule {}